- Now curcmd->argc is filled correctly
- Adapted the new id naming to the debugger : mapped files should not
also load their dependences since this is done already by the linkmap
which is more powerful than doing with a preset variable like done in
elfsh.

- Added use of the breakpoint mutex in the thread starting and thread
creating functions, so that a thread cannot be created or cannot
be started while another thread is doing a sensitive operation like
breakpointing. This allow the debugger to know the exact number of
threads during the treatment of a breakpoint, to avoid manipulating
an inconsistant hash table of threads.

NOTE: The debugger is still NOT WORKING PROPERLY in that tree, dont try
to do anything with it.

thats all

-maym