Corrected some bugs and "finished" porting mjollnir to SPARC
'analyse' runs well until it hits mjr_function_flow_parents_save(), where it tries to copy some data from the _real_ address of the function (ie. the address where it would be if the binary was loaded)
that's a bug? maybe the function intended to copy data from the buffer storing the file contents, or just to copy the address itself?