Summary: Added callgraph walk and reflection on mjrlink_t (0.81a15)
- Added callgraph walk in evarista so that we translate all 
functions after the other in calling order.

- Added reflection on mjrlink_t

- Added a field in mjrlink_t to indicate the scope. Scope
can be local or global. A local scope for a bloc link indicate
a jump in the same function. A global scope for a link indicate
a call or ret. A local scope for function indicate a function
call within the same object. A global scope for function indicates
an external function call (via PLT). This scope field allows to
stop recursion in the callgraph walk and the cfg walk.

- Added link type and link scope constants within the eresi language